ahora el timer va
